Stop adding frames to backtrace once a buttercup wrap func is found
Buttercup wrapped functions are the only thing containing the
buttercup--mark-stackframe marker anyway.

The assert in `buttercup--enclosed-expr' had to be removed, since that
function is also used to verify if the argument is a wrapper function.
Once it is used to test objects that are not wrapped functions, the
assert becomes a liability.